Posted on Twitter!
I finally was able to post on twitter using the IntegrationTest class I
added. This ones uses test resource filtering to plug in the values from
an individual developers .ms2/settings.xml file into the
logback-test.xml

I also found out I had to disable the Twitter4J logger by explicitly
setting it to the NullLoggerFactory. The reason is that it would find
slf4j on the classpath and then that would start a chain reaction that
would ultimately end in failure.